Display color image and upgrade NuGet package to 3.0.0.
