+++ SimpleViewer.Android/SimpleViewer.Android.csproj
@@ -52,7 +52,7 @@
       <Private>True</Private>
     </Reference>
     <Reference Include="Dicom.Core, Version=3.0.0.0, Culture=neutral, PublicKeyToken=3a13f649e28eb09a, processorArchitecture=MSIL">
-      <HintPath>..\packages\fo-dicom.Android.3.0.0-rc3\lib\MonoAndroid1\Dicom.Core.dll</HintPath>
+      <HintPath>..\packages\fo-dicom.Android.3.0.0\lib\MonoAndroid1\Dicom.Core.dll</HintPath>
       <Private>True</Private>
     </Reference>
     <Reference Include="Mono.Android" />
