+++ SimpleViewer.Android/packages.config
@@ -1,6 +1,6 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
   <package id="CSJ2K" version="2.0.0.1" targetFramework="monoandroid60" />
-  <package id="fo-dicom.Android" version="3.0.0-rc3" targetFramework="monoandroid60" />
+  <package id="fo-dicom.Android" version="3.0.0" targetFramework="monoandroid60" />
   <package id="Portable.LibJpeg.NET" version="1.5.1.1" targetFramework="monoandroid60" />
 </packages>
